Chenged output to byte in java and kernels to squeeza the memory (kernels are modulo 2 so we don't need a full int to store results)

+++ src/main/java/it/cvdlab/lar/model/OutputVectorsContainer.java
@@ -10,7 +10,7 @@ import org.slf4j.LoggerFactory;
 public class OutputVectorsContainer {
 
 	@JsonProperty("lista_vettori")
-	private List<List<Integer>> vectorList;
+	private List<List<Byte>> vectorList;
 
 	@JsonProperty("lista_offset")
 	private List<List<Integer>> vectorOffset;
@@ -23,11 +23,11 @@ public class OutputVectorsContainer {
 	
 	public OutputVectorsContainer() {}	
 
-	public List<List<Integer>> getVectorList() {
+	public List<List<Byte>> getVectorList() {
 		return vectorList;
 	}
 
-	public void setVectorList(List<List<Integer>> vectorList) {
+	public void setVectorList(List<List<Byte>> vectorList) {
 		this.vectorList = vectorList;
 	}
